backend/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
defusedxml==0.5.0
Django==2.1
django-allauth==0.37.1
django-appconf==1.0.2
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-select2==6.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
kombu==4.2.1
matplotlib==2.1.2
numpy==1.14.0
oauthlib==2.1.0
opencv-python==3.4.0.12
opencv-contrib-python==3.4.0.12
Pillow==5.0.0
pip-autoremove==0.9.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
vine==1.1.4
